requirements.txt
gensim==3.8.1
seaborn==0.10.0
pandas==1.0.3
scipy==1.2.1
nltk==3.4.5
so_processing==0.1
matplotlib==3.1.3
numpy==1.18.2
tqdm==4.45.0
requests==2.22.0
natsort==7.0.1
progressbar2==3.50.0
python_igraph==0.8.0
scikit_learn==0.22.2.post1
fastparquet==0.3.3
lxml==4.5.0
